+++ 19-1-2023/src/main/java/com/PBL3/controllers/admin/business/BusinessController.java
@@ -1,10 +1,11 @@
-package com.PBL3.controllers.admin.api.business;
+package com.PBL3.controllers.admin.business;
 
-import com.PBL3.dtos.BusinessTypesDTO;
-import com.PBL3.services.IBusinessTypesService;
+import com.PBL3.dtos.BusinessDTO;
+import com.PBL3.services.IBusinessService;
 import com.PBL3.utils.Constants.Constants;
 import com.PBL3.utils.exceptions.ErrorHandler;
 import com.PBL3.utils.helpers.Helper;
+import com.PBL3.utils.response.Message;
 
 import javax.inject.Inject;
 import javax.servlet.ServletException;
@@ -15,18 +16,18 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/business/type"})
+@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/business"})
 @MultipartConfig
-public class BusinessTypesController extends HttpServlet {
+public class BusinessController extends HttpServlet {
 
     private static final long serialVersionUID = 1L;
     @Inject
-    private IBusinessTypesService businessTypesService;
+    private IBusinessService businessService;
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        BusinessTypesDTO businessTypeDto = Helper.paramsToString(req.getParameterMap()).toModel(BusinessTypesDTO.class);
-        ErrorHandler.handle(resp, () -> businessTypesService.createBusinessType(businessTypeDto));
+        BusinessDTO businessDto = Helper.paramsToString(req.getParameterMap()).toModel(BusinessDTO.class);
+        Message message = businessService.createBusiness(businessDto);
+        ErrorHandler.handle(resp, () -> businessService.createBusiness(businessDto));
     }
-
 }
