+++ 19-1-2023/src/main/java/com/PBL3/controllers/admin/plan/PlansController.java
@@ -1,24 +1,26 @@
-package com.PBL3.controllers.admin.api.report;
+package com.PBL3.controllers.admin.plan;
 
-import com.PBL3.services.IReportService;
+import com.PBL3.services.IPlanService;
 import com.PBL3.utils.Constants.Constants;
 import com.PBL3.utils.exceptions.ErrorHandler;
 
 import javax.inject.Inject;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.MultipartConfig;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/reports"})
-public class ReportsController extends HttpServlet {
+@MultipartConfig
+@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/plans"})
+public class PlansController extends HttpServlet {
     @Inject
-    private IReportService iReportService;
+    private IPlanService iPlanService;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        ErrorHandler.handle(resp, () -> iReportService.findAll());
+        ErrorHandler.handle(resp, () -> iPlanService.getAll());
     }
 }
