+++ 19-1-2023/src/main/java/com/PBL3/controllers/admin/planInspectors/PlanInspectorsController.java
@@ -1,7 +1,7 @@
-package com.PBL3.controllers.admin.api.productsCertificates;
+package com.PBL3.controllers.admin.planInspectors;
 
-import com.PBL3.dtos.ProductCertificateDTO;
-import com.PBL3.services.IProductCertificateServie;
+import com.PBL3.dtos.PlanInspectorDTO;
+import com.PBL3.services.IPlanInspectorService;
 import com.PBL3.utils.Constants.Constants;
 import com.PBL3.utils.exceptions.ErrorHandler;
 import com.PBL3.utils.helpers.Helper;
@@ -15,20 +15,21 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/product/certificate"})
+@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + Constants.ADMIN + "/plan/inspectors"})
 @MultipartConfig
-public class ProductsCertificatesController extends HttpServlet {
+public class PlanInspectorsController extends HttpServlet {
     @Inject
-    private IProductCertificateServie iProductCertificateServie;
+    private IPlanInspectorService iPlanInspectorService;
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        ProductCertificateDTO dto = Helper.paramsToString(req.getParameterMap()).toModel(ProductCertificateDTO.class);
-        ErrorHandler.handle(resp, () -> iProductCertificateServie.createOne(dto));
+        PlanInspectorDTO dto = Helper.paramsToString(req.getParameterMap()).toModel(PlanInspectorDTO.class);
+        ErrorHandler.handle(resp, () -> iPlanInspectorService.createOne(dto));
     }
 
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        ErrorHandler.handle(resp, () -> iProductCertificateServie.deleteOne(req.getParameter("id")));
+        PlanInspectorDTO dto = Helper.paramsToString(req.getParameterMap()).toModel(PlanInspectorDTO.class);
+        ErrorHandler.handle(resp, () -> iPlanInspectorService.deactive(dto));
     }
 }
