+++ 19-1-2023/src/main/java/com/PBL3/controllers/admin/product/ProductController.java
@@ -1,4 +1,4 @@
-package com.PBL3.controllers.admin.api.product;
+package com.PBL3.controllers.admin.product;
 
 import com.PBL3.dtos.ProductDTO;
 import com.PBL3.services.IProductService;
