+++ 19-1-2023/src/main/java/com/PBL3/controllers/admin/product/ProductsController.java
@@ -1,26 +1,24 @@
-package com.PBL3.controllers.admin.api.plan;
+package com.PBL3.controllers.admin.product;
 
-import com.PBL3.services.IPlanService;
+import com.PBL3.services.IProductService;
 import com.PBL3.utils.Constants.Constants;
 import com.PBL3.utils.exceptions.ErrorHandler;
 
 import javax.inject.Inject;
 import javax.servlet.ServletException;
-import javax.servlet.annotation.MultipartConfig;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@MultipartConfig
-@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/plans"})
-public class PlansController extends HttpServlet {
+@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/products"})
+public class ProductsController extends HttpServlet {
     @Inject
-    private IPlanService iPlanService;
+    private IProductService iProductService;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        ErrorHandler.handle(resp, () -> iPlanService.getAll());
+        ErrorHandler.handle(resp, () -> iProductService.getAllProducts());
     }
 }
