+++ 19-1-2023/src/main/java/com/PBL3/controllers/admin/report/ReportController.java
@@ -1,0 +1,29 @@
+package com.PBL3.controllers.admin.report;
+
+import com.PBL3.dtos.ReportDTO;
+import com.PBL3.services.IReportService;
+import com.PBL3.utils.Constants.Constants;
+import com.PBL3.utils.exceptions.ErrorHandler;
+import com.PBL3.utils.helpers.Helper;
+
+import javax.inject.Inject;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.MultipartConfig;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet(urlPatterns = {Constants.URL_V1 + Constants.PRIVATE + "/report"})
+@MultipartConfig
+public class ReportController extends HttpServlet {
+    @Inject
+    private IReportService iReportService;
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        ReportDTO dto = Helper.paramsToString(req.getParameterMap()).toModel(ReportDTO.class);
+        ErrorHandler.handle(resp, () -> iReportService.createOne(dto));
+    }
+}
